Базовая реалізація MVVM с асінхронностью

+++ app/src/main/java/com/shukevich/recycleview/screens/UserListViewModel.kt
@@ -3,22 +3,49 @@ package com.shukevich.recycleview.screens
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
+import com.shukevich.recycleview.R
+import com.shukevich.recycleview.UserActionListener
 import com.shukevich.recycleview.model.User
 import com.shukevich.recycleview.model.UserService
 import com.shukevich.recycleview.model.UsersListener
+import com.shukevich.recycleview.tasks.*
 
-class UserListViewModel(
+
+data class UserListItem(
+    val user: User,
+    val isInProgress: Boolean
+)
+
+class UsersListViewModel(
     private val usersService: UserService
-):ViewModel() {
+) : BaseViewModel(), UserActionListener {
 
-    private val _users = MutableLiveData<List<User>>()
-    val users: LiveData<List<User>> = _users
+    private val _users = MutableLiveData<Result<List<UserListItem>>>()
+    val users: LiveData<Result<List<UserListItem>>> = _users
+
+    private val _actionShowDetails = MutableLiveData<Event<User>>()
+    val actionShowDetails: LiveData<Event<User>> = _actionShowDetails
+
+    private val _actionShowToast = MutableLiveData<Event<Int>>()
+    val actionShowToast: LiveData<Event<Int>> = _actionShowToast
+
+    private val userIdsInProgress = mutableSetOf<Long>()
+    private var usersResult: Result<List<User>> = EmptyResult()
+        set(value) {
+            field = value
+            notifyUpdates()
+        }
 
     private val listener: UsersListener = {
-        _users.value = it
+        usersResult = if (it.isEmpty()) {
+            EmptyResult()
+        } else {
+            SuccessResult(it)
+        }
     }
 
     init {
+        usersService.addListener(listener)
         loadUsers()
     }
 
@@ -27,17 +54,66 @@ class UserListViewModel(
         usersService.removeListener(listener)
     }
 
-    fun loadUsers(){
-        usersService.addListener(listener)
+    fun loadUsers() {
+        usersResult = PendingResult()
+        usersService.loadUsers()
+            .onError {
+                usersResult = ErrorResult(it)
+            }
+            .autoCancel()
     }
 
-    fun moveUser(user: User, moveBy: Int){
+    override fun onUserMove(user: User, moveBy: Int) {
+        if (isInProgress(user)) return
+        addProgressTo(user)
         usersService.moveUser(user, moveBy)
+            .onSuccess {
+                removeProgressFrom(user)
+            }
+            .onError {
+                removeProgressFrom(user)
+                _actionShowToast.value = Event(R.string.cant_move_user)
+            }
+            .autoCancel()
     }
 
-    fun deleteUser(user: User) {
+    override fun onUserDelete(user: User) {
+        if (isInProgress(user)) return
+        addProgressTo(user)
         usersService.deleteUser(user)
+            .onSuccess {
+                removeProgressFrom(user)
+            }
+            .onError {
+                removeProgressFrom(user)
+                _actionShowToast.value = Event(R.string.cant_delete_user)
+            }
+            .autoCancel()
     }
 
+    override fun onUserDetails(user: User) {
+        _actionShowDetails.value = Event(user)
+    }
+
+    private fun addProgressTo(user: User) {
+        userIdsInProgress.add(user.id)
+        notifyUpdates()
+    }
+
+    private fun removeProgressFrom(user: User) {
+        userIdsInProgress.remove(user.id)
+        notifyUpdates()
+    }
+
+    private fun isInProgress(user: User): Boolean {
+        return userIdsInProgress.contains(user.id)
+    }
+
+    private fun notifyUpdates() {
+        _users.postValue(usersResult.map { users ->
+            users.map { user -> UserListItem(user, isInProgress(user)) }
+        })
+    }
 
 }
+
